Outsourced classes to inv_items
