+++ pokete.py
@@ -22,8 +22,9 @@ import pokete_data as p_data
 from pokete_classes.color import Color
 from pokete_classes.effects import effects
 from pokete_classes.ui_elements import StdFrame2, Box, ChooseBox, InfoBox, InputBox
-from pokete_classes.classes import PlayMap, Types, InvItem, LearnDisc, Settings, OutP, ResizeScreen, LoadingScreen
+from pokete_classes.classes import PlayMap, Types, Settings, OutP, ResizeScreen, LoadingScreen
 from pokete_classes.health_bar import HealthBar
+from pokete_classes.inv_items import InvItem, LearnDisc
 from pokete_general_use_fns import *
 from release import *
